+++ include/ctranslate2/decoding.h
@@ -1,34 +1,37 @@
 #pragma once
 
 #include "ctranslate2/layers/decoder.h"
+#include "ctranslate2/sampling.h"
 
 namespace ctranslate2 {
 
   void greedy_search(layers::Decoder& decoder,
                      layers::DecoderState& state,
+                     const Sampler& sampler,
                      StorageView& sample_from,
                      StorageView& candidates,
                      const StorageView& memory,
                      const StorageView& memory_lengths,
-                     size_t start_step,
-                     size_t end_token,
-                     size_t max_length,
-                     size_t min_length,
+                     dim_t start_step,
+                     dim_t end_token,
+                     dim_t max_length,
+                     dim_t min_length,
                      std::vector<std::vector<std::vector<size_t>>>& sampled_ids,
                      std::vector<std::vector<float>>& scores,
                      std::vector<std::vector<std::vector<std::vector<float>>>>* attention = nullptr);
 
   void beam_search(layers::Decoder& decoder,
                    layers::DecoderState& state,
+                   const Sampler& sampler,
                    StorageView& sample_from,
                    StorageView& candidates,
                    const StorageView& memory,
                    const StorageView& memory_lengths,
-                   size_t start_step,
-                   size_t end_token,
-                   size_t max_length,
-                   size_t min_length,
-                   size_t beam_size,
+                   dim_t start_step,
+                   dim_t end_token,
+                   dim_t max_length,
+                   dim_t min_length,
+                   dim_t beam_size,
                    size_t num_hypotheses,
                    float length_penalty,
                    std::vector<std::vector<std::vector<size_t>>>& sampled_ids,