+++ src/storage_view.cc
@@ -3,6 +3,14 @@
 #include "./device_dispatch.h"
 
 #define PRINT_MAX_VALUES 6
+#define GUARD_DIM(DIM, RANK)                                            \
+  do {                                                                  \
+    if (DIM >= RANK)                                                    \
+      THROW_INVALID_ARGUMENT("can't index dimension "                   \
+                             + std::to_string(DIM)                      \
+                             + " for a storage with rank "              \
+                             + std::to_string(RANK));                   \
+  } while (false)
 
 namespace ctranslate2 {
 
@@ -52,8 +60,8 @@ namespace ctranslate2 {
     return _dtype;
   }
 
-  size_t StorageView::reserved_memory() const {
-    size_t buffer_size = 0;
+  dim_t StorageView::reserved_memory() const {
+    dim_t buffer_size = 0;
     TYPE_DISPATCH(_dtype, buffer_size = _allocated_size * sizeof (T));
     return buffer_size;
   }
@@ -73,9 +81,11 @@ namespace ctranslate2 {
     return clear();
   }
 
-  StorageView& StorageView::reserve(size_t size) {
+  StorageView& StorageView::reserve(dim_t size) {
+    if (size <= _allocated_size)
+      return *this;
     release();
-    size_t required_bytes = 0;
+    dim_t required_bytes = 0;
     TYPE_DISPATCH(_dtype, required_bytes = size * sizeof (T));
     DEVICE_DISPATCH(_device, _data = primitives<D>::alloc_data(required_bytes));
     if (_data == nullptr)
@@ -85,7 +95,7 @@ namespace ctranslate2 {
     return *this;
   }
 
-  size_t StorageView::rank() const {
+  dim_t StorageView::rank() const {
     return _shape.size();
   }
 
@@ -93,24 +103,26 @@ namespace ctranslate2 {
     return _shape;
   }
 
-  size_t StorageView::dim(ssize_t dim) const {
+  dim_t StorageView::dim(dim_t dim) const {
     if (dim < 0)
       dim = _shape.size() + dim;
+    GUARD_DIM(dim, rank());
     return _shape[dim];
   }
 
-  size_t StorageView::stride(ssize_t dim) const {
+  dim_t StorageView::stride(dim_t dim) const {
     if (dim < 0)
       dim = _shape.size() + dim;
-    return stride(_shape, dim);
+    GUARD_DIM(dim, rank());
+    return compute_stride(_shape, dim);
   }
 
-  size_t StorageView::size() const {
+  dim_t StorageView::size() const {
     return _size;
   }
 
   bool StorageView::is_scalar() const {
-    return rank() == 1 && _size == 1;
+    return _size == 1 && _shape.empty();
   }
 
   bool StorageView::empty() const {
@@ -118,38 +130,42 @@ namespace ctranslate2 {
   }
 
   StorageView& StorageView::reshape(const Shape& new_shape) {
-    if (_size != size(new_shape))
-      THROW_INVALID_ARGUMENT("new shape size (" + std::to_string(size(new_shape))
+    const dim_t new_size = compute_size(new_shape);
+    if (_size != new_size)
+      THROW_INVALID_ARGUMENT("new shape size (" + std::to_string(new_size)
                              + ") is incompatible with current size (" + std::to_string(_size) + ")");
     _shape = new_shape;
     return *this;
   }
 
   StorageView& StorageView::resize(const Shape& new_shape) {
-    if (new_shape.empty())
-      return clear();
-    size_t new_size = size(new_shape);
-    if (new_size > _allocated_size)
-      reserve(new_size);
+    const dim_t new_size = compute_size(new_shape);
+    reserve(new_size);
     _size = new_size;
-    return reshape(new_shape);
+    _shape = new_shape;
+    return *this;
   }
 
-  StorageView& StorageView::resize(size_t dim, size_t new_size) {
+  StorageView& StorageView::resize(dim_t dim, dim_t new_size) {
+    GUARD_DIM(dim, rank());
     Shape new_shape(_shape);
     new_shape[dim] = new_size;
     return resize(new_shape);
   }
 
-  StorageView& StorageView::grow(size_t dim, size_t size) {
+  StorageView& StorageView::grow(dim_t dim, dim_t size) {
+    GUARD_DIM(dim, rank());
     return resize(dim, _shape[dim] + size);
   }
 
-  StorageView& StorageView::shrink(size_t dim, size_t size) {
+  StorageView& StorageView::shrink(dim_t dim, dim_t size) {
+    GUARD_DIM(dim, rank());
     return resize(dim, _shape[dim] - size);
   }
 
   StorageView& StorageView::resize_as(const StorageView& other) {
+    if (other.empty())
+      return clear();
     return resize(other.shape());
   }
 
@@ -197,7 +213,7 @@ namespace ctranslate2 {
   }
 
   template <typename T>
-  T StorageView::scalar_at(const std::vector<size_t>& indices) const {
+  T StorageView::scalar_at(const std::vector<dim_t>& indices) const {
     T scalar = T();
     DEVICE_DISPATCH(_device, scalar = primitives<D>::deref(index<T>(indices), 0));
     return scalar;
@@ -211,7 +227,7 @@ namespace ctranslate2 {
   }
 
   template <typename T>
-  StorageView& StorageView::copy_from(const T* data, size_t size, Device device) {
+  StorageView& StorageView::copy_from(const T* data, dim_t size, Device device) {
     ASSERT_DTYPE(DataTypeToEnum<T>::value);
     if (size != _size)
       THROW_INVALID_ARGUMENT("buffer to copy is of size " + std::to_string(size)
@@ -230,20 +246,17 @@ namespace ctranslate2 {
     return *this;
   }
 
-  size_t StorageView::size(const Shape& shape) {
-    if (shape.empty())
-      return 0;
-    size_t size = 1;
-    for (auto dim : shape)
+  dim_t StorageView::compute_size(const Shape& shape) {
+    dim_t size = 1;
+    for (const dim_t dim : shape)
       size *= dim;
     return size;
   }
 
-  size_t StorageView::stride(const Shape& shape, size_t dim) {
-    if (shape.empty())
-      return 0;
-    size_t stride = 1;
-    for (size_t i = shape.size() - 1; i > dim; --i)
+  dim_t StorageView::compute_stride(const Shape& shape, dim_t dim) {
+    GUARD_DIM(dim, static_cast<dim_t>(shape.size()));
+    dim_t stride = 1;
+    for (dim_t i = shape.size() - 1; i > dim; --i)
       stride *= shape[i];
     return stride;
   }
@@ -267,18 +280,18 @@ namespace ctranslate2 {
       printable.dtype(),
       const auto* values = printable.data<T>();
       if (printable.size() <= PRINT_MAX_VALUES) {
-        for (size_t i = 0; i < printable.size(); ++i) {
+        for (dim_t i = 0; i < printable.size(); ++i) {
           os << ' ';
           print_value(os, values[i]);
         }
       }
       else {
-        for (size_t i = 0; i < PRINT_MAX_VALUES / 2; ++i) {
+        for (dim_t i = 0; i < PRINT_MAX_VALUES / 2; ++i) {
           os << ' ';
           print_value(os, values[i]);
         }
         os << " ...";
-        for (size_t i = printable.size() - (PRINT_MAX_VALUES / 2); i < printable.size(); ++i) {
+        for (dim_t i = printable.size() - (PRINT_MAX_VALUES / 2); i < printable.size(); ++i) {
           os << ' ';
           print_value(os, values[i]);
         }
@@ -286,10 +299,14 @@ namespace ctranslate2 {
       os << std::endl);
     os << "[" << device_to_str(storage.device())
        << " " << dtype_name(storage.dtype()) << " storage viewed as ";
-    for (size_t i = 0; i < storage.rank(); ++i) {
-      if (i > 0)
-        os << 'x';
-      os << storage.dim(i);
+    if (storage.is_scalar())
+      os << "scalar";
+    else {
+      for (dim_t i = 0; i < storage.rank(); ++i) {
+        if (i > 0)
+          os << 'x';
+        os << storage.dim(i);
+      }
     }
     os << ']';
     return os;
@@ -307,10 +324,10 @@ namespace ctranslate2 {
 
 #define DECLARE_IMPL(T)                                                 \
   template T                                                            \
-  StorageView::scalar_at(const std::vector<size_t>& indices) const;     \
+  StorageView::scalar_at(const std::vector<dim_t>& indices) const;      \
   template StorageView& StorageView::fill(T value);                     \
   template StorageView&                                                 \
-  StorageView::copy_from(const T* data, size_t size, Device device);
+  StorageView::copy_from(const T* data, dim_t size, Device device);
 
   DECLARE_ALL_TYPES(DECLARE_IMPL)