+++ CHANGELOG.md
@@ -4,6 +4,30 @@
 
 ### Fixes and improvements
 
+## [v1.3.0](https://github.com/OpenNMT/CTranslate2/releases/tag/v1.3.0) (2020-01-14)
+
+### New features
+
+* Support random sampling (see the `sampling_topk` and `sampling_temperature` translation options)
+* `CT2_CUDA_CACHING_ALLOCATOR_CONFIG` environment variable to configure the CUDA caching allocator
+
+### Fixes and improvements
+
+* Fix incorrect translations on Windows due to incompatibility between the compiler OpenMP and Intel OpenMP
+* Release cuDNN/cuBLAS/TensorRT handles on thread exit when destroying a `TranslatorPool`
+* Remove use of `--{start,end}-group` compiler options when compiling on Mac OS
+* Update Intel MKL to 2020.0 in Docker images
+* Load vocabulary assets for SavedModel exported with OpenNMT-tf 2.5 and above
+
+## [v1.2.3](https://github.com/OpenNMT/CTranslate2/releases/tag/v1.2.3) (2019-12-11)
+
+### Fixes and improvements
+
+* Improve translator robustness on empty batch and inputs
+* Speed optimization for `LayerNorm`
+* Check vocabulary size when converting OpenNMT-tf models
+* Add more samples in the execution profiling output which now supports nested functions
+
 ## [v1.2.2](https://github.com/OpenNMT/CTranslate2/releases/tag/v1.2.2) (2019-11-25)
 
 ### Fixes and improvements